requirements.txt
# RunPod SDK
runpod>=1.7.0

# Chatterbox TTS
chatterbox-tts>=0.1.0

# Audio processing
torchaudio>=2.1.0
soundfile>=0.12.0

# Utilities
numpy>=1.24.0
